OwnerController and StartupController tests

+++ src/main/java/cz/cvut/fit/tjv/poberboh/server/controller/OwnerController.java
@@ -33,6 +33,10 @@ public class OwnerController {
 
     @DeleteMapping("/{id}")
     public void delete(@PathVariable Integer id) throws NotFoundException {
-        ownerService.delete(id);
+        if (ownerService.readById(id).isEmpty()) {
+            throw new NotFoundException("Owner not found");
+        } else {
+            ownerService.delete(id);
+        }
     }
 }
